+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+### 2016-12-09
+* add Included Files Panel
+
 ### 2016-11-29
 * removed from dependencies: 
   - slim/twig-view